PR 301 Suggestions, moving around functionality
