+++ Basic-Car-Maintenance/Shared/Utilities/TextFieldClearButton.swift
@@ -1,0 +1,38 @@
+//
+//  TextFieldClearButton.swift
+//  Basic-Car-Maintenance
+//
+//  https://github.com/mikaelacaron/Basic-Car-Maintenance
+//  See LICENSE for license information.
+//
+
+import SwiftUI
+
+struct TextFieldClearButton: ViewModifier {
+    @Binding var text: String
+    
+    func body(content: Content) -> some View {
+        content
+            .overlay {
+                if !text.isEmpty {
+                    HStack {
+                        Spacer()
+                        Button {
+                            text = ""
+                        } label: {
+                            Image(systemName: "multiply.circle.fill")
+                                .imageScale(.medium)
+                        }
+                        .foregroundColor(.secondary)
+                        .padding(.trailing, 4)
+                    }
+                }
+            }
+    }
+}
+
+extension View {
+    func showClearButton(_ text: Binding<String>) -> some View {
+        self.modifier(TextFieldClearButton(text: text))
+    }
+}
